Memory leak fix in Gecko and Webkit based browsers. An action would be registered to the keyup event, but it would never get unregistered because when calling removeEventListener it was called with "onkeyup" as argument instead of "keyup".
Memory leak fix and annoying bug fix.
When the hotkey dialog is displayed it registers a hotkey for enter and a
hotkey for escape. If there already existed a hotkey for those keys those
wouldn't get unregistered which means the keydown and keyup events would
never get freed and we have a memory leak. The fix goes once step further
and saves the previously registered hotkeys to enter and escape and when
the hotkey dialog is hidden it reregisters the previosly registered
hotkeys. This fixes a bug on the hybrid page where you would navigate
the table with the tab-key and then a hotkey would be registered to enter
to open an errand but if you when you have a errand selected would perform
another hotkey the hotkey dialog would be displayed and will you get the
idea.