Fix for Ckeditor input being reset.
The issue was that textValue() in the TextView component sets the text-value state.
The text-value state is the default state. That means that updateFormValue() is going to be called.
The function since we are using a Ckeditor will call setData() on the Ckeditor. When Ckeditor sets
the data the whole iframe is recreated. This can take time. Rougly 100 ms in Firefox 1.5 at the moment
on my Linux installation. If another textValue() function call is made during those 100 ms Ckeditor
will return empty data. Since the textValue() call will result in an updateFormValue() call Ckeditor
will be set with empty data.

Since the data is being fetched either from the textarea node or Ckeditor it seems uneeded to update
either the textarea node or Ckeditor with the data again so what it does now is that it updates
the text-value state but it does not do it by calling setState() so this whole problem is avoided.