Fix issue reported by Rebtel where content pasted in Firefox caused strange characters to appear in plain text answers.
This was because the Wysiwyg editor paste did not handle the content in any way so it pasted the full HTML.
The full HTML was from Word or other editor causing many strange tags to be inserted. Some of these tags
themselfs had content and when those tags were stripped (and not their content) the strange content was left
inside the answer.